Only remove spec directories when they are on top-level

+++ lib/cleanup_vendor.rb
@@ -3,37 +3,82 @@ require 'cleanup_vendor/version'
 module CleanupVendor
   class Error < StandardError; end
 
-  DIRECTORIES = %w[.git .github spec]
-  EXTENSIONS = %w[c cpp gem h hpp java log md mk o rdoc txt]
-  FILENAMES = %w[README Makefile LICENSE CHANGELOG .dockerignore .gitignore .rspec .travis.yml .yardopts]
+  DEFAULTS = {
+    extensions: %w[c cpp gem h hpp java log md mk o rdoc txt],
+    filenames: %w[README Makefile LICENSE CHANGELOG .dockerignore .gitignore .rspec .travis.yml .yardopts],
+    top_level_directories: %w[spec],
+    directories: %w[.git .github]
+  }.freeze
 
-  def self.cleanup(directory, directories: DIRECTORIES, extensions: EXTENSIONS, filenames: FILENAMES)
-    dir = directory || 'vendor/bundle/ruby'
+  class << self
+    def run(dir, opts = {})
+      summary = []
 
-    filter(dir, directories: directories, extensions: extensions, filenames: filenames) do |f|
-      FileUtils.remove_entry(f)
+      filter(dir, {}.merge(DEFAULTS, opts)) do |f|
+        if opts[:summary]
+          files = File.file?(f) ? [f] : dir_entries(f).grep_v('.').map { |file| File.join(f, file) }
+          summary << files.map { |f| File.stat(File.join(dir, f)) }
+        end
+
+        puts "Removing #{f}..." if opts[:dry_run] || opts[:verbose]
+        FileUtils.remove_entry(f) unless opts[:dry_run]
+      end
+
+      if opts[:summary]
+        all_files = summary.flatten
+        count = all_files.count
+        blocks = all_files.map(&:blocks).sum
+        bytes = all_files.map(&:size).sum
+
+        puts 'Summary:'
+        puts format_summary('Removed files:', count)
+        puts format_summary('Total blocks:', blocks)
+        puts format_summary('Total bytes:', bytes)
+      end
     end
-  end
 
-  def self.filter(dir, directories: [], extensions: [], filenames: [], &block)
-    raise Error.new('Not a directory') unless dir && File.directory?(dir)
+    def filter(dir, opts = {})
+      raise Error.new('Not a directory') unless dir && File.directory?(dir)
 
-    return to_enum(:filter, dir, directories: directories, extensions: extensions, filenames: filenames) unless block_given?
+      unless block_given?
+        return to_enum(:filter, dir, opts)
+      end
 
-    Dir.chdir(dir)
+      extensions = opts[:extensions] || []
+      filenames = opts[:filenames] || []
+      directories = opts[:directories] || []
+      top_level_directories = opts[:top_level_directories] || []
 
-    Dir.glob('**/*') do |f|
-      basename = File.basename(f)
+      dir_entries(dir) do |f|
+        basename = File.basename(f)
 
-      if File.file?(f)
-        if filenames.include?(basename) || extensions.include?(File.extname(basename).delete('.'))
-          yield(f)
+        if File.file?(f)
+          if filenames.include?(basename) || extensions.include?(File.extname(basename).delete('.'))
+            yield(f)
+          end
+        end
+
+        if File.directory?(f)
+          if directories.include?(basename)
+            yield(f)
+          elsif top_level_directories.include?(basename)
+            tld = File.join(dir, File.dirname(f))
+            yield(f) unless dir_entries(tld, '*.gemspec').empty?
+          end
         end
       end
+    end
 
-      if File.directory?(f) && directories.include?(basename)
-        yield(f)
+    def dir_entries(dir, pattern = '**/*', &block)
+      Dir.chdir(dir) do
+        Dir.glob(pattern, File::FNM_DOTMATCH, &block)
       end
     end
+    private :dir_entries
+
+    def format_summary(prefix, number)
+      "\t#{prefix}\t#{number.to_s.rjust(20)}"
+    end
+    private :format_summary
   end
 end
